Refactor: Replace modal with inline question editing
